scripts/testray_automation_tasks/requirements.txt
certifi==2025.8.3
chardet~=5.2.0
charset-normalizer==3.4.3
defusedxml==0.7.1
distro~=1.9.0
idna==3.10
jira>=3.10.1
jellyfish~=1.2.0
oauthlib==3.3.1
olefile~=0.47
packaging==25.0
pip~=25.2
pillow==11.3.0
pycryptodome==3.23.0
PySocks~=1.7.1
python-dateutil~=2.9.0
python-dotenv~=1.1.1
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
sentence-transformers~=5.1.0
setuptools~=80.9.0
tabulate==0.9.0
typing_extensions==4.15.0
urllib3==2.5.0
